services/push_service/requirements.txt
fastapi
uvicorn
pydantic
aio_pika
httpx
python-dotenv